% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angleSink.R
\name{angleSink}
\alias{angleSink}
\title{Calculate a sink angle matrix based on the Point of Interest
Based on Afán et al. 2015}
\usage{
angleSink(POI = c(145.2, -38.8), lat, lon)
}
\arguments{
\item{POI}{central reference point (eg: colony location)}

\item{lat}{latitudes to compare (dimensions must match lon)}

\item{lon}{longitudes to compare (dimensions must match lat)}
}
\value{
A numeric array, with dimensions matching water.direction and water.speed
}
\description{
Calculate a sink angle matrix based on the Point of Interest
Based on Afán et al. 2015
}
\keyword{cost}
\keyword{currents}
\keyword{sink}

