% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCurrents.R
\name{convertCurrents}
\alias{convertCurrents}
\title{Convert current from .nc file to speed and direction values.}
\usage{
convertCurrents(zonal.current, meridional.current)
}
\arguments{
\item{zonal.current}{Extracted using ncvar_get}

\item{meridional.current}{Extracted using ncvar_get}
}
\value{
A list containing 2 elements, "direction" and "speed"
}
\description{
Convert current from .nc file to speed and direction values.
}
\keyword{current}
\keyword{direction}
\keyword{speed}
