% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coresToUse.R
\name{coresToUse}
\alias{coresToUse}
\title{coresToUse, a function for guessing at a good number of cores to use in parallel processing,
base on the computer being used}
\usage{
coresToUse()
}
\value{
integer value of number of cores to be used
}
\description{
coresToUse, a function for guessing at a good number of cores to use in parallel processing,
base on the computer being used
}

