% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costCalcMaster.R
\name{costCalcMaster}
\alias{costCalcMaster}
\title{This function takes a datafile of ocean currents, and uses it to process a dynamic area of influence.}
\usage{
costCalcMaster(currents.file, lat.variableName = NA, lon.variableName = NA,
  depth.variableName = NA, time.variableName = NA,
  current.variableNames = NA, depth.touse = NULL, depth.dimension = NULL,
  dates.range = NA, buffer.days = 10L, output.folder, POI,
  foraging.distance, cell.size = 0.1, split.quant = 0.75,
  circuitscape.link = ifelse(Sys.info()["sysname"] == "Windows",
  "C:/\\"Program Files\\"/Circuitscape/cs_run.exe", "python2.7"), parallel)
}
\arguments{
\item{currents.file}{complete link to .nc data file containing ocean currents data}

\item{lat.variableName}{character name of latitude variable in .nc currents.file. Default "lat"}

\item{lon.variableName}{character name of longitude variable in .nc currents.file. Default "lon"}

\item{depth.variableName}{character name of depth variable in .nc currents.file. Default "depth"}

\item{time.variableName}{character name of time variable in .nc currents.file. Default "time"}

\item{current.variableNames}{character vector (length 2), names of current variables in .nc currents.file.
Element 1 is for the zonal current values, element 2 is for the meridional current values.
Default c( "ZonalCurrent", "MeridionalCurrent" )}

\item{depth.touse}{numerical depth value. Which available depth dataset (if more than one) is to be analysed here?}

\item{depth.dimension}{integer value specifying which dimension of the currents arrays in the .nc file represents
"depth". This will be used to subset only the depth to be analysed here. depth.dimension is treated in the same way
as a MARGIN value, for example a value of 3 represents x[,,i].}

\item{dates.range}{range of dates to subset from input data. If NA, all dates in input file are used.}

\item{buffer.days}{buffer around requested date to search for data}

\item{output.folder}{complete link to a folder to use for temporary files during processing.
A temporary folder will be created within this specified folder, and deleted when computation has completed.}

\item{POI}{c( [longitude], [latitude] ) marking the central point (eg: colony)}

\item{foraging.distance}{A distance in km to consider as the area of interest from the POI.}

\item{cell.size}{Resolution (in degrees) to create for output. Note data will be interpolated from input.}

\item{split.quant}{The point at which to split for areas of influence (0.75 takes the TOP 25\% influence), as per Afán et al 2015}

\item{circuitscape.link}{complete link to executable file to be used for running Circuitscape.
Under Windows, this can be a `.exe` file for Circuitscape itself. Under *nix systems, this can be a link to python,
under which Circuitscape will be called.}

\item{parallel}{FALSE will not use multi-threading, TRUE will guess at an optimal number of cores based on processor,
integer value specifies a number of cores to use when multi-threading tasks.}
}
\description{
The process used is as per Afán et al 2015
}
\keyword{chlorophyll,}
\keyword{circuitscape}
\keyword{currents,}
\keyword{ocean,}

