% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costFun.R
\name{costFun}
\alias{costFun}
\title{Calculate a "cost sink" based on direction, speed, and angle.
Based on Afán et al. 2015}
\usage{
costFun(water.direction, water.speed, angle.sink)
}
\arguments{
\item{water.direction}{A numeric array expressing wind directions.}

\item{water.speed}{An array with dimensions matching that of water.direction}

\item{angle.sink}{A matrix of angles calculated using angleSink.}
}
\value{
A numeric array, with dimensions matching water.direction and water.speed
}
\description{
Calculate a "cost sink" based on direction, speed, and angle.
Based on Afán et al. 2015
}
\keyword{cost}
\keyword{sink}
\keyword{wind}
