% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2Ascii.R
\name{df2Ascii}
\alias{df2Ascii}
\title{df2Ascii function to convert the columns of a data frame to Ascii files, ready for further processing outside of R}
\usage{
df2Ascii(lat, lon, data, file, noDataValue = -9999)
}
\arguments{
\item{lat}{vector or matrix of latitude points}

\item{lon}{vector or matrix of longitude points}

\item{data}{vector or matrix, with dimensions matching lat and lon}

\item{file}{path to write output}

\item{noDataValue}{value to be used in Ascii file as representing a "no data" point}
}
\value{
data.frame
}
\description{
df2Ascii function to convert the columns of a data frame to Ascii files, ready for further processing outside of R
}

