% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFromNc.R
\name{extractFromNc}
\alias{extractFromNc}
\title{extractFromNc get data from a .nc file}
\usage{
extractFromNc(file, data.variable)
}
\arguments{
\item{file}{.nc file containing chlorophyll levels for the entire area of influence}

\item{data.variable}{character string corresponding to the relevant data subset within the .nc file (defaults to "chlorophyll")}
}
\value{
A data.table, containing variables for date and chlorophyll level
}
\description{
extractFromNc get data from a .nc file
}
\keyword{file}
\keyword{import}
\keyword{nc}

