% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{idDub}
\alias{idDub}
\title{calculate data based on inverse distance weighted interpolation}
\usage{
idDub(i, inputlat, inputlon, inputdata, outputlat, outputlon, landmask)
}
\arguments{
\item{i}{row number to analyse}

\item{inputlat}{latitudes of input data}

\item{inputlon}{longitudes of input data}

\item{inputdata}{input data}

\item{outputlat}{latitudes of output data}

\item{outputlon}{longitudes of output data}

\item{landmask}{aligned data for masking}
}
\description{
calculate data based on inverse distance weighted interpolation
}
\keyword{average}
\keyword{distance}
\keyword{inverse}
\keyword{weighted}

