% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landMask.R
\name{landMask}
\alias{landMask}
\title{landMask function to download a local map, and use it to create a mask of the local land areas}
\usage{
landMask(lat, lon, cores = TRUE)
}
\arguments{
\item{lat}{vector or matrix of latitude points}

\item{lon}{vector or matrix of longitude points}

\item{cores}{number of cores to use when multi-threading}
}
\value{
data.frame
}
\description{
landMask function to download a local map, and use it to create a mask of the local land areas
}

