% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapDistance.R
\name{latKmToDeg}
\alias{latKmToDeg}
\title{Convert either latitude or longitude distances between angles (degrees) and distances (km)}
\usage{
latKmToDeg(km)
}
\arguments{
\item{km}{Numeric value, distance in km}
}
\value{
numeric value
}
\description{
Convert either latitude or longitude distances between angles (degrees) and distances (km)
}
\keyword{gis}
\keyword{maps}

