% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapDistance.R
\name{lonDegToKm}
\alias{lonDegToKm}
\title{Convert either latitude or longitude distances between angles (degrees) and distances (km)}
\usage{
lonDegToKm(degrees, lat = 39.32633)
}
\arguments{
\item{degrees}{Numeric value, degrees for distance}

\item{lat}{Numeric value, latitude position}
}
\value{
numeric value
}
\description{
Convert either latitude or longitude distances between angles (degrees) and distances (km)
}
\keyword{gis}
\keyword{maps}

