% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapDistance.R
\name{lonKmToDeg}
\alias{lonKmToDeg}
\title{Convert either latitude or longitude distances between angles (degrees) and distances (km)}
\usage{
lonKmToDeg(km, lat = 39.32633)
}
\arguments{
\item{km}{Numeric value, distance in km}

\item{lat}{Numeric value, latitude position}
}
\value{
numeric value
}
\description{
Convert either latitude or longitude distances between angles (degrees) and distances (km)
}
\keyword{gis}
\keyword{maps}

