% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapDistance.R
\name{mapDistance}
\alias{mapDistance}
\title{Take two latlon coordinates, and return the distance between them.}
\usage{
mapDistance(lat.1, lon.1, lat.2, lon.2, unit = "km")
}
\arguments{
\item{lat.1}{latitude position, point 1}

\item{lon.1}{longitude position, point 1}

\item{lat.2}{latitude position, point 2}

\item{lon.2}{longitude position, point 2}

\item{unit}{distance units desired for output, default "km"}
}
\value{
numeric value, km between the two points
}
\description{
Take two latlon coordinates, and return the distance between them.
}
\keyword{gis}
\keyword{maps}

