% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixRotate.R
\name{matrixRotate}
\alias{matrixRotate}
\title{matrixRotate}
\usage{
matrixRotate(x, angle = c(0, 90, 180, 270))
}
\arguments{
\item{x}{a matrix}

\item{angle}{numeric, multiple of 90, positive (clockwise) or negative (anti-clockwise)}
}
\description{
matrixRotate
}

