% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanChlorophyll.R
\name{meanChlorophyll}
\alias{meanChlorophyll}
\title{meanChlorophyll Calculating the relevant metric for a single time}
\usage{
meanChlorophyll(date, data.chlorophyll, data.conductance, max.day.diff = 31)
}
\arguments{
\item{date}{date for which to calculate the chlorophyll level}

\item{data.chlorophyll}{data extracted from .nc file}

\item{data.conductance}{data calculated using `costCalcMaster`, based on currents}

\item{max.day.diff}{integer, number of days to allow as a difference between a conductance array
and the corresponding data array}
}
\value{
numeric value for mean chlorophyll
}
\description{
meanChlorophyll Calculating the relevant metric for a single time
}

