% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navto.R
\name{navto}
\alias{navto}
\title{A function to help navigate through folders
will create folders requested if they don't exist already}
\usage{
navto(folder.names)
}
\arguments{
\item{folder.names}{A vector of folder through which to navigate}
}
\description{
A function to help navigate through folders
will create folders requested if they don't exist already
}
\keyword{folders}
