% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_variableNames.R
\name{nc_variableNames}
\alias{nc_variableNames}
\title{nc_variableNames
a simple function for viewing the names of the variables within a .nc file}
\usage{
nc_variableNames(file)
}
\arguments{
\item{file}{character string linking to .nc file}
}
\value{
a character vector of variable names
}
\description{
nc_variableNames
a simple function for viewing the names of the variables within a .nc file
}

