% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotButterfly.R
\name{plotButterfly}
\alias{plotButterfly}
\title{plotButterfly View a "butterfly"}
\usage{
plotButterfly(conductance.table, which = 1)
}
\arguments{
\item{conductance.table}{as output from costCalcMaster}

\item{which}{either an integer or a date (character string), which date to plot}
}
\value{
TRUE (silently)
}
\description{
plotButterfly View a "butterfly"
}
\keyword{plot}
\keyword{view}
