% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processConductance.R
\name{processConductance}
\alias{processConductance}
\title{processConductance convert files output from circuitscape processing into a data table}
\usage{
processConductance(conductance.file, land.mask, split.quant = 0.75,
  file.identifier = "_dynamic", write.file = TRUE, return.dt = FALSE,
  return.data.vector = FALSE)
}
\arguments{
\item{conductance.file}{file to be processed}

\item{land.mask}{object to specify which values to be masked as "land"}

\item{split.quant}{numeric value, at what fraction should the dataset be split?}

\item{file.identifier}{character string identifying files}

\item{write.file}{TRUE/FALSE should the file be written to disk?}

\item{return.dt}{TRUE/FALSE should the output be converted to a data.table?}

\item{return.data.vector}{TRUE/FALSE should the output be returned as a vector?}
}
\value{
none
}
\description{
processConductance convert files output from circuitscape processing into a data table
}
\keyword{conductance,}
\keyword{currents}
\keyword{ocean}
