% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeIsPositive.R
\name{slopeIsPositive}
\alias{slopeIsPositive}
\title{slopeIsPositive check whether the slope of a vector is positive}
\usage{
slopeIsPositive(x)
}
\arguments{
\item{x}{a vector}
}
\value{
boolean TRUE if slope is positive, FALSE if negative or zero
}
\description{
slopeIsPositive check whether the slope of a vector is positive
}
