% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chlorophyllCalc.R
\name{chlorophyllCalc}
\alias{chlorophyllCalc}
\title{chlorophyllCalc Use the relevant local area of influence (calculated using the butterfly algorithm) to find chlorophyll levels over time}
\usage{
chlorophyllCalc(input.file, data.variable = "chlorophyll", conductance.data,
  max.day.diff = 31, exclude.nas = TRUE)
}
\arguments{
\item{input.file}{.nc file containing chlorophyll levels for the entire area of influence}

\item{data.variable}{character string corresponding to the relevant data subset within the .nc file (defaults to "chlorophyll")}

\item{conductance.data}{input for calculation}

\item{max.day.diff}{integer, number of days to allow as a difference between a conductance array}

\item{exclude.nas}{should NA values be excluded?}
}
\value{
A data.table, containing variables for date and chlorophyll level
}
\description{
chlorophyllCalc Use the relevant local area of influence (calculated using the butterfly algorithm) to find chlorophyll levels over time
}
\keyword{chlorophyll}
\keyword{cost}
\keyword{currents}

