% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCircuitscape.R
\name{runCircuitscape}
\alias{runCircuitscape}
\title{runCircuitscape a function for externally applying the circuitscape algorithms to give data}
\usage{
runCircuitscape(i, costs.filelist, csrun.link, output.folder, source.asc.link,
  sink.asc.link, dates)
}
\arguments{
\item{i}{integer, used to iterate over multiple files}

\item{costs.filelist}{list of ascii files to be processed}

\item{csrun.link}{path to python file to be used to run circuitscape}

\item{output.folder}{where to put the processed data}

\item{source.asc.link}{"source" file for cuircuitscape}

\item{sink.asc.link}{"sink" file for cuircuitscape}

\item{dates}{list of dates corresponding to time series datapoints}

\item{python.call}{specify how to call python}
}
\value{
none
}
\description{
runCircuitscape a function for externally applying the circuitscape algorithms to give data
}
\keyword{ascii}
\keyword{circuitscape,}

